requirements.txt
amqp==1.4.9
anyjson==0.3.3
arrow==0.11.0
asn1crypto==0.23.0
bcrypt==3.1.4
billiard==3.3.0.23
celery==3.1.23
cffi==1.11.2
cryptography==2.1.3
Django==1.11.7
django-ckeditor==5.3.1
django-constance==2.0.0
django-extensions==1.9.7
django-js-asset==0.1.1
django-picklefield==1.0.0
django-rest-framework==0.1.0
djangorestframework==3.3.3
Fabric3==1.13.1.post1
future==0.16.0
idna==2.6
isort==4.2.15
Jinja2==2.8
jinja2-time==0.1.0
kik==1.0.7
kombu==3.0.37
MarkupSafe==1.0
messengerbot==0.1.4
paramiko==2.4.0
permabots==2.2.3
pyasn1==0.3.7
pycparser==2.18
PyNaCl==1.2.0
python-dateutil==2.6.1
python-telegram-bot==4.2.0
pytz==2017.3
redis==2.10.6
requests==2.10.0
shortuuid==0.4.3
six==1.10.0
typing==3.6.2
